old/source/requirements.txt
matplotlib==3.10.7
numpy==2.3.5
planetary_computer==1.0.0
protobuf==6.33.2
psycopg2_binary==2.9.11
pyproj==3.7.2
pystac==1.14.1
pystac_client==0.9.0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
Requests==2.32.5
rioxarray==0.19.0
source==1.2.0
